Rework to use DeeplinkRedirectLoadingIndicator
