+++ src/components/DeeplinkWrapper/DeeplinkRedirectLoadingIndicator.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {View} from 'react-native';
+import {withOnyx} from 'react-native-onyx';
 import TextLink from '../TextLink';
 import Text from '../Text';
 import Icon from '../Icon';
@@ -9,14 +10,29 @@ import * as Expensicons from '../Icon/Expensicons';
 import colors from '../../styles/colors';
 import styles from '../../styles/styles';
 import withLocalize, {withLocalizePropTypes} from '../withLocalize';
+import Navigation from '../../libs/Navigation/Navigation';
+import ROUTES from '../../ROUTES';
+import compose from '../../libs/compose';
+import ONYXKEYS from '../../ONYXKEYS';
 
 const propTypes = {
     openLinkInBrowser: PropTypes.func.isRequired,
 
+    session: PropTypes.shape({
+        /** Currently logged-in user email */
+        email: PropTypes.string,
+    }),
+
     ...withLocalizePropTypes,
 };
 
-function DeeplinkRedirectLoadingIndicator(props) {
+const defaultProps = {
+    session: {
+        email: '',
+    },
+};
+
+function DeeplinkRedirectLoadingIndicator({translate, openLinkInBrowser, session}) {
     return (
         <View style={styles.deeplinkWrapperContainer}>
             <View style={styles.deeplinkWrapperMessage}>
@@ -27,11 +43,12 @@ function DeeplinkRedirectLoadingIndicator(props) {
                         src={Illustrations.RocketBlue}
                     />
                 </View>
-                <Text style={[styles.textHeadline, styles.textXXLarge]}>{props.translate('deeplinkWrapper.launching')}</Text>
+                <Text style={[styles.textHeadline, styles.textXXLarge]}>{translate('deeplinkWrapper.launching')}</Text>
                 <View style={[styles.mt2, styles.fontSizeNormal, styles.textAlignCenter]}>
-                    <Text>{props.translate('deeplinkWrapper.redirectedToDesktopApp')}</Text>
+                    <Text>{translate('deeplinkWrapper.loggedInAs', {email: session.email})}</Text>
                     <Text style={[styles.textAlignCenter]}>
-                        {props.translate('deeplinkWrapper.youCanAlso')} <TextLink onPress={props.openLinkInBrowser}>{props.translate('deeplinkWrapper.openLinkInBrowser')}</TextLink>.
+                        {translate('deeplinkWrapper.doNotSeePrompt')} <TextLink onPress={() => openLinkInBrowser(true)}>{translate('deeplinkWrapper.tryAgain')}</TextLink>
+                        {translate('deeplinkWrapper.or')} <TextLink onPress={() => Navigation.navigate(ROUTES.HOME)}>{translate('deeplinkWrapper.continueInWeb')}</TextLink>.
                     </Text>
                 </View>
             </View>
@@ -48,6 +65,14 @@ function DeeplinkRedirectLoadingIndicator(props) {
 }
 
 DeeplinkRedirectLoadingIndicator.propTypes = propTypes;
+DeeplinkRedirectLoadingIndicator.defaultProps = defaultProps;
 DeeplinkRedirectLoadingIndicator.displayName = 'DeeplinkRedirectLoadingIndicator';
 
-export default withLocalize(DeeplinkRedirectLoadingIndicator);
+export default compose(
+    withLocalize,
+    withOnyx({
+        session: {
+            key: ONYXKEYS.SESSION,
+        },
+    }),
+)(DeeplinkRedirectLoadingIndicator);
